[ADD] Validações se houver erros

+++ index.js
@@ -4,10 +4,16 @@ const fs = require('fs');
 const MoodleController = require('./src/moodleController');
 
 var config;
+const RUN_WITHOUT_CLIENT = true;
+
+function main() {
+    if (!verifyConfig()) return;
+
+    runClient();
+}
 
 function verifyConfig() {
     if (!fs.existsSync('config.json')) {
-
         fs.writeFileSync('config.json', JSON.stringify(
             {
                 ra: "RA",
@@ -25,52 +31,72 @@ function verifyConfig() {
 }
 
 function runClient() {
-    const client = new Client({
-        puppeteer: {
-            headless: true, // false - Visible | true - Invisible
-            args: ['--no-sandbox']
-        },
-        authStrategy: new LocalAuth()
-    });
+    if (!RUN_WITHOUT_CLIENT) {
+        const client = new Client({
+            puppeteer: {
+                headless: true, // false - Visible | true - Invisible
+                args: ['--no-sandbox']
+            },
+            authStrategy: new LocalAuth()
+        });
+    
+        client.on('qr', qr => {
+            qrcode.generate(qr, { small: true });
+        });
+    
+        client.on('ready', async () => {
+            console.log('Cliente Whatsapp Iniciado.');
+            getNextLessons(client);
+    
+        });
+    
+        client.initialize();
+        console.log("Iniciando cliente...");
+    } else {
+        getNextLessons(null);
+    }
+    
+}
 
-    client.on('qr', qr => {
-        qrcode.generate(qr, { small: true });
-    });
+async function getNextLessons(client) {
+    const moodleController = new MoodleController(config.ra, config.password);
 
-    client.on('ready', async () => {
-        console.log('Cliente Iniciado');
+    console.log("Iniciando Requisição ao Moodle...")
+    let lessons = await moodleController.main();
 
-        const moodleController = new MoodleController(config.ra, config.password);
-
-        console.log("Iniciando Moodle...")
-        let lessons = await moodleController.main();
-
-        let msg = createMsg(lessons);
-
-        // Check if chat exists
-        let chats = await client.getChats();
-        chats = chats.filter(chat => { return chat.name == config.chatName && chat.isGroup; });
-        if (chats.length == 1) {
-            await client.sendMessage(chats[0].id._serialized, msg);
-            console.log("Mensagem enviada!");
-        } else {
-            console.log("Chat não encontrado!");
+    if (!moodleController.connected) {
+        if (!RUN_WITHOUT_CLIENT) await client.sendMessage(config.chatName, "Ocorreu um erro ao tentar conectar ao Moodle.");
+    } else {
+        console.log("Requisição finalizada!");
+        if (!moodleController.gotLessons) {
+            console.log("Ocorreu um erro ao tentar buscar as lições.");
+            if (!RUN_WITHOUT_CLIENT) await client.sendMessage(config.chatName, "Ocorreu um erro ao tentar buscar as lições.");
         }
-
+        else if (lessons.length == 0) {
+            console.log("Nenhuma lição encontrada!");
+            if (!RUN_WITHOUT_CLIENT) await client.sendMessage(config.chatName, "Nenhuma lição encontrada!");
+        }
+        else {
+            let msg = createMsg(lessons);
+            console.log("Mensagem: \n" + msg + "\n");
+            if (!RUN_WITHOUT_CLIENT) await sendMsg(client, msg);
+        }
+        
+    }
+    
+    if (!RUN_WITHOUT_CLIENT) {
         console.log("Finalizando cliente em 5 segundos...");
         setTimeout(() => client.destroy(), 5000);
-
-    });
-
-    client.initialize();
-    console.log("Iniciando cliente...");
+    }
+    
 }
 
 function createMsg(lessons) {
-    let msg = "Boa Noite.";
+    let msg = "Próximas lições:";
 
     let lessonsGroup = {
     };
+    
     lessons.forEach((lesson) => {
         if (lessonsGroup[lesson.course] == undefined) lessonsGroup[lesson.course] = [];
 
@@ -97,33 +123,21 @@ function createMsg(lessons) {
     return msg;
 }
 
-async function runWithoutWpp() {
-    const moodleController = new MoodleController(config.ra, config.password);
-    let licoes = await moodleController.main();
-    console.log(licoes);
-}
+async function sendMsg(client, msg) {
+    try {
+        let chats = await client.getChats();
+        chats = chats.filter(chat => { return chat.name == config.chatName && chat.isGroup; });
 
-function main() {
-    if (!verifyConfig()) return;
-
-    runClient();
-    //runWithoutWpp();
+        if (chats.length == 0) console.log("Grupo não encontrado!");
+        else if (chats.length > 1) console.log("Mais de um grupo encontrado!");
+        else if (chats.length == 1) {
+            console.log("Grupo encontrado!");
+            await client.sendMessage(chats[0].id._serialized, msg);
+            console.log("Mensagem enviada!");
+        }
+    } catch (e) {
+        client.sendMessage(config.chatName, "Ocorreu um erro ao tentar enviar a mensagem.");
+    }
 }
 
 main();
-
-
-
-
-
-/* 
-    Boa Noite.
-    As Atividades dos próximos 3 dias são:
-        -**• Atividade**: Design de Interação - Desk Research (**05/04 00:00**)
-        -**• Atividade**: Design de Interação - Pesquisa de Usabilidade (**06/04 00:00**)
-        -**• Atividade**: Design de Interação - Prototipação (**06/04 00:00**)
-        -**• Atividade**: Design de Interação - Teste de Usabilidade (**06/04 00:00**)
-        -**• Atividade**: Design de Interação - Entrevista (**06/04 00:00**)
-    
-
-*/
